# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/webhooks/clerk": {
      "post": {
        "operationId": "ClerkController_handleClerkWebhook",
        "parameters": [
          {
            "name": "svix-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "svix-timestamp",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "svix-signature",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "tags": ["Clerk"]
      }
    },
...
    "/user/product/{productId}": {
      "get": {
        "operationId": "UserController_getUsersByProductId",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["user"]
      }
    },
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProductDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Products"]
      },
      "get": {
        "operationId": "ProductsController_findAllProductsForThisUser",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Products"]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Products"]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProductDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Products"]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Products"]
      }
    },
...
  },
  "info": {
    "title": "Pineapple API",
    "description": "The Pineapple API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateProductDto": { "type": "object", "properties": {} },
      "UpdateProductDto": { "type": "object", "properties": {} },
...
    }
  }
}